fix: switch to mise, global balena, ipv4 key pulls
GPG was using IPv6 even though I don't have it on my host, so had to
force IPv4 in some awful ways.

The `mise` balena-cli registry doesn't work properly because of how
`balena` access native modules. So, we had to switch to brew for this.
Even my `asdf` plugin wasn't working. I don't care enough to try to fix
that, so, here we are.

Switched to `.mise.toml` to reflect that's all I'm using/supporting now.